Refactor Fitbit API Controllers and Restructure Packages
Remove FitbitController and split responsibilities into new ProfileController and WeightController. Relocate DTO and converter classes into specific profile, weight, and sleep subpackages for better organization and maintainability.
