+++ fitness-service/src/main/kotlin/com/michibaum/fitness_service/fitbit/api/weight/WeightController.kt
@@ -1,0 +1,30 @@
+package com.michibaum.fitness_service.fitbit.api.weight
+
+import com.michibaum.authentication_library.security.jwt.JwtAuthentication
+import com.michibaum.fitness_service.api.weight.WeightService
+import com.michibaum.fitness_service.fitbit.FitbitOAuth
+import com.michibaum.fitness_service.fitbit.api.FitbitApi
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RestController(value = "fitbitWeightController")
+class WeightController(
+    val fitbitApi: FitbitApi,
+    val fitbitOAuth: FitbitOAuth,
+    val fitbitWeightConverter: FitbitWeightConverter,
+    val weightService: WeightService,
+) {
+
+    @PostMapping(value = ["/api/fitbit/weight/{startDate}/{endDate}"])
+    fun updateWeight(@PathVariable startDate: String, @PathVariable endDate: String, principal: JwtAuthentication){
+        val credentials = fitbitOAuth.getCredentials(principal)
+        credentials?.let {
+            val weight = fitbitApi.weightLog(it, startDate, endDate).map { weight ->
+                fitbitWeightConverter.toDomain(weight, principal.getUserId())
+            }
+            weightService.update(weight)
+        }
+    }
+
+}
