+++ fitness-service/src/main/kotlin/com/michibaum/fitness_service/fitbit/api/weight/WeightLogDto.kt
@@ -1,4 +1,4 @@
-package com.michibaum.fitness_service.fitbit.api
+package com.michibaum.fitness_service.fitbit.api.weight
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
