+++ gateway-service/src/main/resources/application.yml
@@ -46,8 +46,6 @@ spring:
           uri: lb://fitness-service
           predicates:
             - Host=fitness.michibaum.ch
-#          filters:
-#            - Authentication  TODO doesnt work because of oAuth
         - id: website
           uri: lb://website-service
           predicates:
